requirements.txt
streamlit
request
pillow